+++ src/index.ts
@@ -9,6 +9,7 @@ export { default as RecaptchaForm} from './components/recaptcha/RecaptchaForm';
 export { default as NotificationSnackbar } from './components/NotificationSnackbar';
 export { default as TextFieldValidate } from './components/TextFieldValidate';
 export { default as TextFieldPassword } from './components/TextFieldPassword';
+export { default as SelectField } from './components/SelectField';
 
 export * from './utils/validateTelefone';
 export * from './utils/validateEmail';